+++ library/image_streamer_golden_image.py
@@ -29,7 +29,7 @@ description:
 version_added: "2.3"
 requirements:
     - "python >= 2.7.9"
-    - "hpOneView >= 3.1.0"
+    - "hpeOneView >= 3.1.0"
 author: "Gustavo Hennig (@GustavoHennig)"
 options:
     state: