+++ library/oneview_event.py
@@ -29,7 +29,7 @@ description:
 version_added: "2.3"
 requirements:
     - "python >= 2.7.9"
-    - "hpOneView >= 3.2.0"
+    - "hpeOneView >= 3.2.0"
 author: "Felipe Bulsoni (@fgbulsoni)"
 options:
     state: