+++ library/oneview_fcoe_network_facts.py
@@ -31,7 +31,7 @@ description:
     - Retrieve the facts about one or more of the FCoE Networks from OneView.
 version_added: "2.4"
 requirements:
-    - hpOneView >= 5.0.0
+    - hpeOneView >= 5.0.0
 author:
     - Felipe Bulsoni (@fgbulsoni)
     - Thiago Miotto (@tmiotto)