+++ library/oneview_firmware_driver.py
@@ -29,7 +29,7 @@ description:
     - Provides an interface to remove Firmware Driver resources.
 requirements:
     - "python >= 2.7.9"
-    - "hpOneView >= 2.0.1"
+    - "hpeOneView >= 2.0.1"
 author: "Bruno Souza (@bsouza)"
 options:
     state: