+++ library/oneview_logical_downlinks_facts.py
@@ -28,7 +28,7 @@ description:
     - Retrieve facts about one or more of the Logical Downlinks from OneView.
 requirements:
     - "python >= 2.7.9"
-    - "hpOneView >= 2.0.1"
+    - "hpeOneView >= 2.0.1"
 author: "Bruno Souza (@bsouza)"
 options:
     name: